Revrt posix timer back to alarm
Using posix timers causes severe performance regression in the hole-null case.
Revert back to alarm() with integer timeout, but keep support for the speical
zero timeout. In this case we expire the timer immediately and the update loop
will exit after the first cycle.

+++ test/benchmark.h
@@ -24,10 +24,10 @@ extern volatile sig_atomic_t timer_is_running;
 
 const char *type_name(enum input_type type);
 int parse_type(const char *name, const char *arg);
-double parse_seconds(const char *name, const char *arg);
+int parse_seconds(const char *name, const char *arg);
 int parse_count(const char *name, const char *arg);
 int64_t parse_size(const char *name, const char *arg);
 
-void start_timer(double seconds);
+void start_timer(int seconds);
 
 #endif /* BENCHMARK_H */
